refactor: align stock model organization with endpoint structure
- Create new model files: filings.rs and ownership.rs for better organization
- Delete alternative.rs, executive.rs, and fund.rs modules
- Redistribute models to align with their endpoint usage:
  - Move MarketStatus to market.rs
  - Move FilingSentiment to sentiment.rs
  - Move IPOCalendar and IPOEvent to models/calendar.rs
  - Move executive/compliance models to compliance.rs
  - Move ownership models to new ownership.rs file
- Update all imports to reflect new model locations
- Improve code maintainability through logical model grouping
