+++ src/models/stock/company.rs
@@ -57,41 +57,3 @@ pub struct Symbol {
     /// Currency.
     pub currency: Option<String>,
 }
-
-/// Supply chain relationship.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SupplyChainRelationship {
-    /// Symbol.
-    pub symbol: Option<String>,
-    /// Company name.
-    pub name: Option<String>,
-    /// Country.
-    pub country: Option<String>,
-    /// 1-tier supplier.
-    #[serde(rename = "oneMonthCorrelation")]
-    pub one_month_correlation: Option<f64>,
-    /// 1-year correlation.
-    #[serde(rename = "oneYearCorrelation")]
-    pub one_year_correlation: Option<f64>,
-    /// 6-month correlation.
-    #[serde(rename = "sixMonthCorrelation")]
-    pub six_month_correlation: Option<f64>,
-    /// 3-month correlation.
-    #[serde(rename = "threeMonthCorrelation")]
-    pub three_month_correlation: Option<f64>,
-    /// 2-week correlation.
-    #[serde(rename = "twoWeekCorrelation")]
-    pub two_week_correlation: Option<f64>,
-    /// 2-year correlation.
-    #[serde(rename = "twoYearCorrelation")]
-    pub two_year_correlation: Option<f64>,
-}
-
-/// Supply chain data.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SupplyChainData {
-    /// Company symbol.
-    pub symbol: String,
-    /// List of suppliers.
-    pub data: Vec<SupplyChainRelationship>,
-}
