+++ src/models/stock/compliance.rs
@@ -150,87 +150,208 @@ pub struct VisaApplication {
     pub h1b_dependent: Option<String>,
 }
 
-/// Filing sentiment analysis.
+/// Supply chain relationship.
 #[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct FilingSentiment {
-    /// Access number.
-    #[serde(rename = "accessNumber")]
-    pub access_number: String,
-    /// Symbol.
-    pub symbol: String,
-    /// CIK.
-    pub cik: String,
-    /// Sentiment scores.
-    pub sentiment: SentimentScores,
-}
-
-/// Sentiment scores.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SentimentScores {
-    /// Percentage of negative words.
-    pub negative: f64,
-    /// Percentage of positive words.
-    pub positive: f64,
-    /// Polarity score.
-    pub polarity: f64,
-    /// Percentage of litigious words.
-    pub litigious: f64,
-    /// Percentage of uncertainty words.
-    pub uncertainty: f64,
-    /// Percentage of constraining words.
-    pub constraining: f64,
-    /// Percentage of modal-weak words.
-    #[serde(rename = "modal-weak")]
-    pub modal_weak: f64,
-    /// Percentage of modal-strong words.
-    #[serde(rename = "modal-strong")]
-    pub modal_strong: f64,
-    /// Percentage of modal-moderate words.
-    #[serde(rename = "modal-moderate")]
-    pub modal_moderate: f64,
-}
-
-/// Document similarity index.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SimilarityIndex {
+pub struct SupplyChainRelationship {
     /// Symbol.
     pub symbol: Option<String>,
-    /// CIK.
-    pub cik: Option<String>,
-    /// Array of similarity data.
-    pub similarity: Vec<SimilarityData>,
+    /// Company name.
+    pub name: Option<String>,
+    /// Country.
+    pub country: Option<String>,
+    /// 1-tier supplier.
+    #[serde(rename = "oneMonthCorrelation")]
+    pub one_month_correlation: Option<f64>,
+    /// 1-year correlation.
+    #[serde(rename = "oneYearCorrelation")]
+    pub one_year_correlation: Option<f64>,
+    /// 6-month correlation.
+    #[serde(rename = "sixMonthCorrelation")]
+    pub six_month_correlation: Option<f64>,
+    /// 3-month correlation.
+    #[serde(rename = "threeMonthCorrelation")]
+    pub three_month_correlation: Option<f64>,
+    /// 2-week correlation.
+    #[serde(rename = "twoWeekCorrelation")]
+    pub two_week_correlation: Option<f64>,
+    /// 2-year correlation.
+    #[serde(rename = "twoYearCorrelation")]
+    pub two_year_correlation: Option<f64>,
 }
 
-/// Similarity data point.
+/// Supply chain data.
 #[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SimilarityData {
-    /// CIK.
-    pub cik: String,
-    /// Access number of the document.
-    #[serde(rename = "accessNumber")]
-    pub access_number: String,
-    /// Item 1 similarity score.
-    pub item1: f64,
-    /// Item 2 similarity score.
-    pub item2: f64,
-    /// Item 1a similarity score.
-    pub item1a: f64,
-    /// Item 7 similarity score.
-    pub item7: f64,
-    /// Item 7a similarity score.
-    pub item7a: f64,
-    /// Form type.
-    pub form: String,
-    /// Report URL.
-    #[serde(rename = "reportUrl")]
-    pub report_url: String,
-    /// Filing URL.
-    #[serde(rename = "filingUrl")]
-    pub filing_url: String,
-    /// Filed date.
-    #[serde(rename = "filedDate")]
-    pub filed_date: String,
-    /// Accepted date.
-    #[serde(rename = "acceptedDate")]
-    pub accepted_date: String,
+pub struct SupplyChainData {
+    /// Company symbol.
+    pub symbol: String,
+    /// List of suppliers.
+    pub data: Vec<SupplyChainRelationship>,
+}
+
+/// Executive or board member information.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Executive {
+    /// Executive name.
+    pub name: Option<String>,
+    /// Age.
+    pub age: Option<i64>,
+    /// Title.
+    pub title: Option<String>,
+    /// Year first appointed.
+    pub since: Option<String>,
+    /// Sex.
+    pub sex: Option<String>,
+    /// Total compensation.
+    pub compensation: Option<i64>,
+    /// Compensation currency.
+    pub currency: Option<String>,
+}
+
+/// Company executives response.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct CompanyExecutives {
+    /// Company symbol.
+    pub symbol: String,
+    /// Array of executives and board members.
+    pub executive: Vec<Executive>,
+}
+
+/// Congressional trading data.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct CongressionalTrade {
+    /// Symbol.
+    pub symbol: String,
+    /// Transaction date.
+    #[serde(rename = "transactionDate")]
+    pub transaction_date: String,
+    /// Transaction amount.
+    #[serde(rename = "transactionAmount")]
+    pub transaction_amount: String,
+    /// Name.
+    pub name: String,
+    /// Owned by.
+    #[serde(rename = "ownedBy")]
+    pub owned_by: String,
+    /// Position.
+    pub position: String,
+    /// Asset name.
+    #[serde(rename = "assetName")]
+    pub asset_name: Option<String>,
+    /// Filing date.
+    #[serde(rename = "filingDate")]
+    pub filing_date: Option<String>,
+}
+
+/// Congressional trading response.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct CongressionalTrading {
+    /// Symbol.
+    pub symbol: String,
+    /// Array of congressional trades.
+    pub data: Vec<CongressionalTrade>,
+}
+
+/// Lobbying data.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct LobbyingData {
+    /// Symbol.
+    pub symbol: String,
+    /// Name.
+    pub name: Option<String>,
+    /// Description.
+    pub description: Option<String>,
+    /// Country.
+    pub country: Option<String>,
+    /// Year.
+    pub year: i32,
+    /// Period.
+    pub period: String,
+    /// Income.
+    pub income: f64,
+    /// Expenses.
+    pub expenses: f64,
+    /// Client ID.
+    #[serde(rename = "clientId")]
+    pub client_id: Option<String>,
+    /// Registrant ID.
+    #[serde(rename = "registrantId")]
+    pub registrant_id: Option<String>,
+}
+
+/// Lobbying response.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Lobbying {
+    /// Symbol.
+    pub symbol: String,
+    /// Array of lobbying data.
+    pub data: Vec<LobbyingData>,
+}
+
+/// USA spending data.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct USASpendingData {
+    /// Symbol.
+    pub symbol: String,
+    /// Recipient name.
+    #[serde(rename = "recipientName")]
+    pub recipient_name: Option<String>,
+    /// Recipient parent name.
+    #[serde(rename = "recipientParentName")]
+    pub recipient_parent_name: Option<String>,
+    /// Award description.
+    #[serde(rename = "awardDescription")]
+    pub award_description: Option<String>,
+    /// Country.
+    pub country: Option<String>,
+    /// Action date.
+    #[serde(rename = "actionDate")]
+    pub action_date: String,
+    /// Total value.
+    #[serde(rename = "totalValue")]
+    pub total_value: f64,
+    /// Performance start date.
+    #[serde(rename = "performanceStartDate")]
+    pub performance_start_date: Option<String>,
+    /// Performance end date.
+    #[serde(rename = "performanceEndDate")]
+    pub performance_end_date: Option<String>,
+    /// Awarding agency name.
+    #[serde(rename = "awardingAgencyName")]
+    pub awarding_agency_name: Option<String>,
+    /// Awarding sub agency name.
+    #[serde(rename = "awardingSubAgencyName")]
+    pub awarding_sub_agency_name: Option<String>,
+    /// Awarding office name.
+    #[serde(rename = "awardingOfficeName")]
+    pub awarding_office_name: Option<String>,
+    /// Performance country.
+    #[serde(rename = "performanceCountry")]
+    pub performance_country: Option<String>,
+    /// Performance city.
+    #[serde(rename = "performanceCity")]
+    pub performance_city: Option<String>,
+    /// Performance county.
+    #[serde(rename = "performanceCounty")]
+    pub performance_county: Option<String>,
+    /// Performance state.
+    #[serde(rename = "performanceState")]
+    pub performance_state: Option<String>,
+    /// Performance zip code.
+    #[serde(rename = "performanceZipCode")]
+    pub performance_zip_code: Option<String>,
+    /// Award type.
+    #[serde(rename = "awardType")]
+    pub award_type: Option<String>,
+    /// NAICS code.
+    #[serde(rename = "naicsCode")]
+    pub naics_code: Option<String>,
+}
+
+/// USA spending response.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct USASpending {
+    /// Symbol.
+    pub symbol: String,
+    /// Array of USA spending data.
+    pub data: Vec<USASpendingData>,
 }
