+++ src/models/stock/insider.rs
@@ -57,26 +57,3 @@ pub struct InsiderSentimentData {
     /// Insider sentiment data.
     pub data: Vec<InsiderSentiment>,
 }
-
-/// Ownership.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Ownership {
-    /// Name.
-    pub name: String,
-    /// Share.
-    pub share: i64,
-    /// Change.
-    pub change: Option<i64>,
-    /// Filing date.
-    #[serde(rename = "filingDate")]
-    pub filing_date: String,
-}
-
-/// Ownership data.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct OwnershipData {
-    /// Symbol.
-    pub symbol: String,
-    /// Ownership data.
-    pub ownership: Vec<Ownership>,
-}
