+++ src/models/stock/mod.rs
@@ -1,34 +1,32 @@
 //! Stock market data models organized by category.
 
-pub mod alternative;
 pub mod analytics;
 pub mod common;
 pub mod company;
 pub mod compliance;
 pub mod corporate_actions;
 pub mod estimates;
-pub mod executive;
+pub mod filings;
 pub mod financials;
-pub mod fund;
 pub mod historical;
 pub mod insider;
 pub mod market;
+pub mod ownership;
 pub mod price;
 pub mod sentiment;
 
 // Re-export all types for backward compatibility
-pub use alternative::*;
 pub use analytics::*;
 pub use common::*;
 pub use company::*;
 pub use compliance::*;
 pub use corporate_actions::*;
 pub use estimates::*;
-pub use executive::*;
+pub use filings::*;
 pub use financials::*;
-pub use fund::*;
 pub use historical::*;
 pub use insider::*;
 pub use market::*;
+pub use ownership::*;
 pub use price::*;
 pub use sentiment::*;
